+++ packages/tinacms/src/components/sidebar/SidebarBody.tsx
@@ -18,6 +18,7 @@ limitations under the License.
 
 import * as React from 'react'
 
+import { Form } from '@tinacms/forms'
 import { useState } from 'react'
 import styled, { keyframes, css, StyledComponent } from 'styled-components'
 import {
@@ -28,10 +29,13 @@ import {
   timing,
   radius,
   shadow,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  ThemeProps,
 } from '@tinacms/styles'
 import { FormList } from './FormList'
 import { useCMS, useSubscribable } from '../../react-tinacms'
-import { FormView, FormHeader } from '../form/FormView'
+import { FormView } from '../form/FormView'
+import { LeftArrowIcon } from '@tinacms/icons'
 
 export const FormsView = () => {
   const [activeFormId, setActiveFormId] = useState<string>()
@@ -92,11 +96,16 @@ export const FormsView = () => {
     <>
       {activeForm && (
         <FormWrapper isEditing={isEditing} isMultiform={isMultiform}>
-          <FormView
-            activeForm={activeForm}
-            setActiveFormId={setActiveFormId}
-            isMultiform={isMultiform}
-          />
+          {isMultiform && (
+            <MultiformFormHeader
+              activeForm={activeForm}
+              setActiveFormId={setActiveFormId}
+            />
+          )}
+          {!isMultiform && activeForm.label && (
+            <FormHeader activeForm={activeForm} />
+          )}
+          <FormView activeForm={activeForm} />
         </FormWrapper>
       )}
     </>
@@ -236,25 +245,6 @@ export const FormBody: StyledComponent<'div', {}, {}> = styled.div`
   }
 `
 
-const FormFooter = styled.div`
-  position: relative;
-  flex: 0 0 auto;
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: 64px;
-  background-color: white;
-  border-top: 1px solid ${color.grey(2)};
-
-  ${Wrapper} {
-    flex: 1 0 auto;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 16px;
-  }
-`
-
 const FormAnimationKeyframes = keyframes`
   0% {
     transform: translate3d( 100%, 0, 0 );
@@ -278,14 +268,14 @@ const FormWrapper = styled.div<FormWrapperProps>`
   width: 100%;
   position: relative;
 
-  ${FormHeader}, ${FormBody}, ${FormFooter} {
+  * > {
     transform: translate3d(100%, 0, 0);
   }
 
   ${p =>
     p.isEditing &&
     css`
-      ${FormHeader}, ${FormBody}, ${FormFooter} {
+      * > {
         transform: none;
         animation-name: ${FormAnimationKeyframes};
         animation-duration: 150ms;
@@ -294,6 +284,103 @@ const FormWrapper = styled.div<FormWrapperProps>`
         animation-timing-function: ease-out;
       }
     `};
+`
+
+export interface MultiformFormHeaderProps {
+  activeForm: Form
+  setActiveFormId(id?: string): void
+}
+
+export const MultiformFormHeader = styled(
+  ({
+    activeForm,
+    setActiveFormId,
+    ...styleProps
+  }: MultiformFormHeaderProps) => {
+    return (
+      <button {...styleProps} onClick={() => setActiveFormId()}>
+        <LeftArrowIcon />
+        <span>{activeForm.label}</span>
+      </button>
+    )
+  }
+)`
+  position: relative;
+  width: 100%;
+  cursor: pointer;
+  border: none;
+  background-image: none;
+  background-color: white;
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  padding: 0 ${padding()} ${padding('small')} ${padding()};
+  color: inherit;
+  font-size: ${font.size(5)};
+  transition: color 250ms ease-out;
+  user-select: none;
+
+  span {
+    flex: 1 1 auto;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    text-align: left;
+  }
+
+  svg {
+    flex: 0 0 auto;
+    width: 24px;
+    fill: ${color.grey(3)};
+    height: auto;
+    transform: translate3d(-4px, 0, 0);
+    transition: transform 150ms ease-out;
+  }
+
+  :hover,
+  :active {
+    color: ${color.primary()};
+    outline: none;
+    border: none;
+
+    svg {
+      fill: ${color.grey(8)};
+      transform: translate3d(-7px, 0, 0);
+      transition: transform 250ms ease;
+    }
+  }
+`
+
+export interface FormHeaderProps {
+  activeForm: Form
+}
+
+export const FormHeader = styled(
+  ({ activeForm, ...styleProps }: FormHeaderProps) => {
+    return (
+      <div {...styleProps}>
+        <span>{activeForm.label}</span>
+      </div>
+    )
+  }
+)`
+  position: relative;
+  width: 100%;
+  background-color: white;
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  padding: 0 ${padding()} ${padding('small')} ${padding()};
+  color: inherit;
+  font-size: ${font.size(5)};
+  user-select: none;
+
+  span {
+    flex: 1 1 auto;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
 `
 
 export const SaveButton: StyledComponent<typeof Button, {}, {}> = styled(