+++ packages/@tinacms/api-git/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/tinacms/tinacms/compare/@tinacms/api-git@0.7.0-alpha.0...@tinacms/api-git@0.7.0) (2020-03-09)
+
+**Note:** Version bump only for package @tinacms/api-git
+
 # [0.7.0-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/api-git@0.5.7-alpha.0...@tinacms/api-git@0.7.0-alpha.0) (2020-03-05)
 
 ### Features