+++ packages/@tinacms/forms/CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/tinacms/tinacms/compare/@tinacms/forms@0.2.1...@tinacms/forms@0.3.0) (2020-03-16)
+
+
+### Features
+
+* expose FormAPI from final-form ([2620f2d](https://github.com/tinacms/tinacms/commit/2620f2d))
+
+
+
+
+
+## [0.2.1](https://github.com/tinacms/tinacms/compare/@tinacms/forms@0.2.1-alpha.0...@tinacms/forms@0.2.1) (2020-03-09)
+
+**Note:** Version bump only for package @tinacms/forms
+
 ## [0.2.1-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/forms@0.2.0...@tinacms/forms@0.2.1-alpha.0) (2020-03-05)
 
 **Note:** Version bump only for package @tinacms/forms